# backend3/docs/swagger.yaml
basePath: /api
definitions:
...
  controller_k8s.CreateSecretRequest:
    properties:
      manifest:
        additionalProperties: true
        type: object
      namespace:
        type: string
    required:
    - manifest
    - namespace
    type: object
...
  controller_k8s.DeleteSecretRequest:
    properties:
      name:
        type: string
      namespace:
        type: string
    required:
    - name
    - namespace
    type: object
...
  controller_k8s.UpdateSecretRequest:
    properties:
      manifest:
        additionalProperties: true
        type: object
      namespace:
        type: string
    required:
    - manifest
    - namespace
    type: object
...
  oneops_backend3_pkg_utils.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8082
info:
  contact: {}
  description: |-
    OneOps 一体化运维平台后端 API
    包含系统管理、CMDB、K8s、审计、监控、应用授权等模块
  title: OneOps API
  version: "1.0"
paths:
...
  /k8s/clusters/{id}/secrets:
    delete:
      consumes:
      - application/json
      description: 删除指定集群中的 Secret
      parameters:
      - description: 集群 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 删除请求(命名空间+名称)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller_k8s.DeleteSecretRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 无效的集群ID / 无权访问 / 删除失败
          schema:
            $ref: '#/definitions/oneops_backend3_pkg_utils.Response'
      security:
      - BearerAuth: []
      summary: 删除 Secret
      tags:
      - K8s-配置
    get:
      description: 分页获取指定集群下指定命名空间的 Secret 列表
      parameters:
      - description: 集群 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 命名空间
        in: query
        name: namespace
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 无效的集群ID / 无权访问 / 获取失败
          schema:
            $ref: '#/definitions/oneops_backend3_pkg_utils.Response'
      security:
      - BearerAuth: []
      summary: 获取 Secret 列表
      tags:
      - K8s-配置
    post:
      consumes:
      - application/json
      description: 在指定集群的命名空间中创建 Secret
      parameters:
      - description: 集群 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 创建请求(含 manifest)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller_k8s.CreateSecretRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 无效的集群ID / 无权访问 / 创建失败
          schema:
            $ref: '#/definitions/oneops_backend3_pkg_utils.Response'
      security:
      - BearerAuth: []
      summary: 创建 Secret
      tags:
      - K8s-配置
    put:
      consumes:
      - application/json
      description: 更新指定集群中的 Secret
      parameters:
      - description: 集群 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新请求(含 manifest)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controller_k8s.UpdateSecretRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 无效的集群ID / 无权访问 / 更新失败
          schema:
            $ref: '#/definitions/oneops_backend3_pkg_utils.Response'
      security:
      - BearerAuth: []
      summary: 更新 Secret
      tags:
      - K8s-配置
  /k8s/clusters/{id}/secrets/{namespace}/{name}:
    get:
      description: 按集群、命名空间、名称获取 Secret 详情
      parameters:
      - description: 集群 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 命名空间
        in: path
        name: namespace
        required: true
        type: string
      - description: Secret 名称
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 无效的集群ID / 无权访问 / 获取失败
          schema:
            $ref: '#/definitions/oneops_backend3_pkg_utils.Response'
      security:
      - BearerAuth: []
      summary: 获取 Secret 详情
      tags:
      - K8s-配置
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 输入 Bearer {token}，token 由 /api/login 获取
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"